% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_from_fits.R
\name{fit_ps_msm}
\alias{fit_ps_msm}
\title{MSM for stratum probabilities}
\usage{
fit_ps_msm(
  object,
  data,
  t0 = object$t0[1],
  which_ps_numerator = "converted",
  which_ps_denominator = c("immune", "doomed", "helped", "helpedplus", "converted",
    "alwaysinf"),
  msm_covariates = "Age",
  msm_formula = "splines::ns(Age, 3)",
  msm_family = binomial(),
  ...
)
}
\arguments{
\item{todo}{TODO}
}
\description{
MSM for stratum probabilities
}
