% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{one_boot}
\alias{one_boot}
\title{Function for one bootstrap replicate}
\usage{
one_boot(
  data,
  asymp_formula,
  symp_formula,
  symp_lr_formula,
  n_boot,
  t0,
  event_name,
  weight_name,
  vax_name,
  bounds
)
}
\arguments{
\item{data}{dataset containing vaccine indicator, time variable, event type variable, and covariates}

\item{asymp_formula}{formula for asymptomatic Cox model}

\item{symp_formula}{formula for symptomatic Cox model}

\item{symp_lr_formula}{formula for LR GLM}

\item{n_boot}{number of bootstrap replicates for bootstrap CI}

\item{t0}{numeric value or numeric vector of t0s}

\item{event_name}{name of event variable}

\item{weight_name}{name of weight variable, if applicable}
}
\value{
dataframe with VE and P estimates
}
\description{
Function for one bootstrap replicate
}
\keyword{internal}
