% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ps_ve}
\alias{print.ps_ve}
\title{Print the output of a \code{"ps_ve"} object.}
\usage{
\method{print}{ps_ve}(object, what = "ve")
}
\arguments{
\item{what}{type to print, default VE estimates, other option probability of belonging to strata}

\item{x}{A \code{"ps_ve"} object.}
}
\description{
Print the output of a \code{"ps_ve"} object.
}
