% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ve_nai}
\alias{print.ve_nai}
\title{Print the output of a \code{"ve_nai"} object.}
\usage{
\method{print}{ve_nai}(object, what = "ve", ...)
}
\arguments{
\item{what}{type to print, default VE estimates, other option probability of belonging to strata}

\item{...}{Other arguments (not used)}

\item{x}{A \code{"ve_nai"} object.}
}
\description{
Print the output of a \code{"ve_nai"} object.
}
