% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_from_fits.R
\name{ve_from_fits}
\alias{ve_from_fits}
\title{Compute VE for naturally asymptomatic infections and other related parameters}
\usage{
ve_from_fits(
  symp_cox_fit,
  asymp_cox_fit,
  symp_lr_fit,
  data,
  t0 = quantile(data$ftime, c(0.2, 0.5, 0.9)),
  ...
)
}
\arguments{
\item{symp_cox_fit}{A \code{coxph} fit for symptomatic infections}

\item{asymp_cox_fit}{A \code{coxph} fit for asymptomatic infections}

\item{symp_lr_fit}{A \code{glm} fit for probability of symptoms given infection}

\item{data}{The \code{data} used to fit the above models. Should have columns \code{ftime}, \code{ftype},
\code{vax}, and other columns for covariates.}

\item{t0}{The time(s) of interest. The code computes cumulative incidence at these prespecified times
and reports VE estimates (and other estimates) at each specified time}
}
\description{
Compute VE for naturally asymptomatic infections and other related parameters
}
