% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_nai.R
\name{ve_nai}
\alias{ve_nai}
\title{Main function for VEnai analysis}
\usage{
ve_nai(
  data,
  vax_name = "vax",
  time_name = "ftime",
  event_name = "ftype",
  symp_ind_name = "symp_ind",
  weight_name = NULL,
  symp_level = 1,
  asymp_level = 2,
  covariate_names = c("Age", "Sex", "HighRiskInd", "EthnicityHispanic",
    "EthnicityNotreported", "EthnicityUnknown", "Black", "Asian", "NatAmer", "PacIsl",
    "Multiracial", "RaceOther", "RaceNotreported", "RaceUnknown"),
  t0 = c(175, 180, 185, 190, 195, 200),
  n_boot = n_boot,
  return_models = FALSE,
  bounds = FALSE,
  sensitivity = FALSE,
  delta = 1,
  seed = 12345
)
}
\arguments{
\item{data}{dataset containing vaccine indicator, time variable, event type variable, and covariates}

\item{vax_name}{name of vaccine variable}

\item{time_name}{name of time of event variable}

\item{event_name}{name of event variable}

\item{symp_ind_name}{name of symptomatic infection indicator variable}

\item{weight_name}{name of weight variable, if applicable}

\item{symp_level}{value taken by symptomatic level of event_name variable}

\item{asymp_level}{value taken by asymptomatic level of event_name variable}

\item{covariate_names}{character vector of covariate names to adjust for}

\item{t0}{numeric value or numeric vector of t0s}

\item{n_boot}{number of bootstrap replicates for bootstrap CI}

\item{return_models}{boolean to return Cox and LR models for estimates, default FALSE}

\item{bounds}{boolean to return bounds on VE_{NAI} without cross-world assumption, default FALSE}

\item{seed}{seed for reproducibility}
}
\value{
VE NAI object containing point estimates, confidence intervals/SE estimates, and (optionally) model fits
}
\description{
Main function for VEnai analysis
}
