% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_nai_bounds.R
\name{ve_nai_bounds}
\alias{ve_nai_bounds}
\title{Function to estimate the bounds on VE_NAI}
\usage{
ve_nai_bounds(
  data,
  vax_name = "vax",
  time_name = "ftime",
  event_name = "ftype",
  symp_level = 1,
  asymp_level = 2,
  t0 = quantile(data[[time_name]], c(0.2, 0.5, 0.9))
)
}
\arguments{
\item{data}{The \code{data} used to fit the above models. Should have columns \code{ftime}, \code{ftype},
\code{vax}, and other columns for covariates.}

\item{vax_name}{name of vaccine variable}

\item{time_name}{name of time of event variable}

\item{event_name}{name of event variable}
}
\description{
Function to estimate the bounds on VE_NAI
}
