% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_nai_sensitivity.R
\name{ve_nai_sensitivity}
\alias{ve_nai_sensitivity}
\title{Function for VE NAI sensitivity analysis}
\usage{
ve_nai_sensitivity(
  symp_cox_fit,
  asymp_cox_fit,
  symp_lr_fit,
  data,
  t0 = quantile(data$ftime, c(0.2, 0.5, 0.9)),
  vax_name = "vax",
  symp_ind_name = "symp_ind",
  delta = 1,
  ...
)
}
\arguments{
\item{tbd}{}
}
\description{
Function for VE NAI sensitivity analysis
}
